Move PathTreeNode pretty printing into pathtreenode
