+++ main/src/core/MonoDevelop.Core/MonoDevelop.FSW/PathTree.cs
@@ -264,34 +264,6 @@ namespace MonoDevelop.FSW
 				next.Previous = node;
 		}
 
-		public override string ToString ()
-		{
-			var sb = new StringBuilder ();
-			PrettyPrint (sb, rootNode, "");
-			return sb.ToString ();
-		}
-
-		void PrettyPrint (StringBuilder builder, PathTreeNode node, string indent)
-		{
-			builder.Append (indent);
-			if (node.Next == null) {
-				builder.Append ("\\-");
-				indent += "  ";
-			} else {
-				builder.Append ("|-");
-				indent += "| ";
-			}
-
-			builder.Append (node.Segment);
-			if (node.IdCount != 0) {
-				builder.AppendFormat (" ({0})", node.IdCount);
-			}
-
-			node = node.FirstChild;
-			while (node != null) {
-				PrettyPrint (builder, node, indent);
-				node = node.Next;
-			}
-		}
+		public string PrettyPrint () => rootNode.ToString ();
 	}
 }
